ball movement works

+++ ball.h
@@ -8,11 +8,12 @@
 
 struct ball_t {
     float x, y, vx, vy;
+    
     ALLEGRO_BITMAP* img;
 };
 typedef struct ball_t ball;
 
 ball* criaBola(float x, float y);
 
-float distance(float x1, float y1, float x2, float y2);
-float vector(float k2, float k1, float dist);
+float distance(float x1, float y1, int x2, int y2);
+float vector(int k2, float k1, float dist);
